Add funções para salvar e imprimir dicionario de distancias
